web: convert mdi-icons script to module
Before (as of e3be77688), I was essentially relying on a hack to
"export" the icons data when dynamically importing mdi-svg.min.js.
Specifically, I was assigning to "window.icons" to make it globally
accessible.

While this works in Chrome, it does not in Safari, so we should do
things the proper way - by explicitly exporting from the module.

concourse/concourse#5131

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>