atc: only bump check order if new versions
Fixes a bug where the check order is bumped everytime it is saved into
the db, regardless if the version already existed. The reason why this
is was probably implemented that way is that lidar no longer knows about
what the "from" version is. I changed it so that the versions are only
bumped if there is at least one new version within the list of versions
returned by the check.

Also the fixed the tests for checking the check order is not bumped with
existing versions, where it was passing because of some missing setup.

[#5127]

Signed-off-by: Clara Fu <cfu@pivotal.io>