atc: change how algorithm deals with version every
Before this change for a version every input, the algorithm would grab
the latest build of a job and try to use the next version after that
build in order to satisfy the version every constraint. This is great
until you have the situation where the latest build used a old pinned
version, and now the "next" version would be the version after that old
pinned version. This will cause the algorithm to produce builds for
versions it has possibly already ran before until it reaches the latest
version.

In order to avoid the possibility of pinning versions to produce builds
using old versions it has already used before, we will now have the
algorithm grab the latest version (rather than build) that has passed
through this job. Then using that latest version that has been used by
the job, it will get the version right after that in order to satisfy
the version every constraint. This added a lot more complexity to the
query, but is the fastest known way to query for that version.

A few indexes were removed from the build inputs table because they are
unused. One was added in order to allow for the new query to run more
efficiently.

[#5125]

Signed-off-by: Clara Fu <cfu@pivotal.io>