atc: change checks gc default to 1m
With the previous default being 6 hours, this caused a lot of build up
of checks within large scale deployments and results in slowness to
the gc query. It was only decided to be a long time period for the
default to allow for operators to find information on the checks that
had previously happened.

We decided to optimize for a faster checks GC rather than more
checks history by default, and if the user deems it necessary to keep
around the history for longer they can configure the GC interval to be
longer.

[#5157]

Signed-off-by: Clara Fu <cfu@pivotal.io>