fly: Remove adminAtcServer
adminAtcServer was added as part of the Super Admin stories. Having a
mock server that returns `is_admin` set to true should not change the
behaviour of fly at all. Therefore we removed it from the entire test
suite.

We updated the releated tests and made sure they made sense in regards
to what functionality they are testing.

Signed-off-by: Taylor Silva <tsilva@pivotal.io>
Co-authored-by: Muntasir Chowdhury <mchowdhury@pivotal.io>
Signed-off-by: Taylor Silva <tsilva@pivotal.io>