containerd: convert garden resource limits to OCI
This commit converts limits for CPU shares, memory, and max number of
pids into the OCI equivalents. This does not include all limits
supported by gdn. In particular, we're missing `CPUQuotaPerShare`,
`TCPMemoryLimit`, `BlockIOWeight`, and the ability to disable the swap
limit. These are hidden behind gdn flags, which will be addressed in
concourse/concourse#5179.

Another limit we are missing is `MaxContainers`. This isn't a
per-container limit, but is also relevant to concourse/concourse#5179.

The testflight integration test for container limits are currently not
passing as we are not setting up cgroups properly yet.

Signed-off-by: Izabela Gomes <igomes@pivotal.io>
Co-authored-by: Aidan Oldershaw <aoldershaw@pivotal.io>