containerd: add cgroup mount and path to OCI spec
We were missing the mount to /sys/fs/cgroup, resulting in that path
being empty on containers. We also need to set the CgroupsPath in the
spec. We followed gdn's implementation, which sets the path to "" for
privileged containers, and "basepath/handle" for non-privileged ones.

We pass `baseCgroupsPath` as an argument to `OciCgroupsPath` because
eventually the base path will need to be more dynamic. Specifically, gdn
changes the base path depending on certain flags
(`enable-cpu-throttling` and `tag`). We will address this in
concourse/concourse#5179.

Signed-off-by: Izabela Gomes <igomes@pivotal.io>
Co-authored-by: Aidan Oldershaw <aoldershaw@pivotal.io>
Signed-off-by: Izabela Gomes <igomes@pivotal.io>