web: structural: track route Transitions
Rather than just giving the SubPage access to the current route, it may
also be useful to give the SubPage access to the route they came from.
In the context of the Build page, this allows us to determine whether
the URL changed in a way that does not require reloading the build info
(e.g. just the highlight changed)

To accomplish this, introduce a `Routes.Transition` type, and add a
`fromBuildPage` field to `Build.Flags` (to keep track of the build page
that we came from, if any).

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>