web: structural: fetch parent and src elements for scroll
Before, we were only fetching the src viewport of the src element. In
order to properly determine the scroll position of the parent when
scrolling to a src element, we'll need the positions of both the src and
parent (note: not viewport of the src) and the viewport of the parent.

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>