web: replace Endpoint `toUrl` with `toPath`
Rather than determining the string form of each endpoint with `toUrl`,
determine the path segments with `toPath`. The motivation for this is to
remove the page information from Endpoints, and move the responsibility
of converting pages to query params to `Api.paginatedGet`.

This is useful since RESTful Endpoints should not imply HTTP verbs. The
`JobBuilds` endpoint, for instance, is not always associated with a GET,
so attaching a `Maybe Page` to it does not make sense in the context of
other verbs - it should not be possible to trigger a build (POST
JobBuilds) passing a `Just Page {...}`

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>