web: introduce per-resource endpoints
Rather than just having all endpoints at the top-level - which resulted
in endpoints being difficult to locate IMO - introduce a new union type
for each resource representing the sub-endpoints that are available for
that resource. For instance, on the Job resource, one sub-endpoint is
PauseJob (corresponding with the `/pause` suffix).

One unfortunate thing is that we can't have overlap in union type
variant names. For instance, both the Job and Pipeline resources have a
`/pause` sub-endpoint, but we can't have a single `Pause` variant - we
need to have `PauseJob` and `PausePipeline`. This results in a bit of
redundancy when specifying endpoints: `PausePipeline |> Pipeline id`,
rather than the preferred `Pause |> Pipeline id`.

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>