web: remove ImageAsset grouping
Initially, I had made a top-level `Asset` type, which had one variant of
`ImageAsset`. This was in an effort to pre-emptively organize the
Assets. However, it seems like nearly every asset that gets referenced
in the Elm code is an image, so moving these to the top-level makes the
code more readable (don't need to pipe to Assets.ImageAsset everywhere)

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>