atc: schedule builds according to algorithm order
Previously, we would schedule pending rerun builds first and then
schedule all other builds but this would cause problems if the job is
serial true. There would be cases when builds would be stuck pending
because it is stuck in max in flight. The situation is when a rerun
build is being scheduled but the max in flight logic would return back
the regular build as the "latest pending build", and that would result
in the rerun build getting stuck at max in flight.

We changed it so that the order that we schedule the builds will be in
the same order as the algorithm, where rerun builds are ordered with the
original builds. For example, the scheduler will order the builds to be
scheduled in order of build IDs ascending: 1, 1.1, 2, rather than before
we would order based on rerun builds first: 1.1, 1, 2.

[#5312]

Co-authored-by: Alex Suraci <suraci.alex@gmail.com>
Signed-off-by: Clara Fu <cfu@pivotal.io>