atc: structure: only assert on pipeline IDs
concourse/concourse#5314

This _should_ result in more robust tests -- the important assertion in
each of these cases seemed to be simply about the set of pipelines that
were visible, and the exact JSON content was a bit of a detail.

Now it should be possible to freely change the JSON structure of those
endpoints without breaking the tests (as long as the "id" field keeps
the same behaviour).

Signed-off-by: Jamie Klassen <cklassen@pivotal.io>
Co-authored-by: Bishoy Youssef <byoussef@pivotal.io>