atc: behaviour: update fly integration tests
concourse/concourse#5329

During the course of working on this issue, we tried to refactor the
existing API tests to be resilient to changes in the JSON structure. It
seems reasonable to have some assertions about the JSON structure of the
output of fly commands using the `--json` flag. Maybe these integration
tests could be similarly refactored to be a bit more robust (maybe
leaving only a single tests that is so obsessed with JSON), but we're
not going to tackle that now.

Signed-off-by: Bishoy Youssef <byoussef@pivotal.io>
Co-authored-by: Jamie Klassen <cklassen@pivotal.io>