web: structure: use Data.pipeline in tests
concourse/concourse#5321

Currently, pipelines are being defined as record literals in many places
throughout the tests. This makes it difficult to, say, add a new field
to the Pipelines type alias.

This commit changes every Pipeline literal to use the Data.pipeline
function. It also adds some helpers for changing the pipeline fields.

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>