hack/jaeger: add elasticsearch backend w kibana
concourse/concourse#5382

We experimented with dynamically generating a data table showing number
of builds and average duration across pipelines. Strictly speaking this
is nothing new, but we could do something similar for specific step
types/names, which would be novel. We didn't come up with a convincing
dataset for this use case, though.

Perhaps the most important insight that got us unstuck today was the
`ES_TAGS_AS_FIELDS_ALL` option in jaeger - this made it possible to
write the queries we needed to get useful graphs in kibana.

Another thing worth noting is that jaeger tends to die if it starts up
before elasticsearch and we found ourselves having to `docker-compose
restart jaeger` when we were starting fresh.

Signed-off-by: Taylor Silva <tsilva@pivotal.io>
Co-authored-by: Jamie Klassen <cklassen@pivotal.io>