web: behaviour: fix scroll on build history list
concourse/concourse#5383

This was broken by concourse/concourse#5275 - in order to get ScrollToId
to work, we needed to replace `getViewportOf` with `getElement`.
However, due to elm/browser#86, the `viewport` and `scene` returned by
`getElement` is of the document rather than the target.

This could be fixed by using `.parentElem` instead of `.srcElem` in the
`getX` and `getY` functions, since we manually update the `viewport` and
`scene` to match the target for `.parentElem`. However, there is a
noticable performance degradation needing to perform 3 browser `Tasks`
(`getElement` * 2 + `getViewportOf` * 1) rather than performing just the
one that's required for the other scroll behaviours. I introduced
`scrollCoordsSimple` to handle this simple case, while leaving
`scrollCoords` alone for the more complex case.

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>