web: behaviour: add dynamically-expanding textarea
concourse/concourse#4781

We decided to use a port for this, since the recommended way for having
a dynamically sized textarea is to set the height to 'auto' and then to
the scrollHeight in succession (deferred by a setTimeout), which is not
easy to achieve in Elm alone.

We also needed to add a second setTimeout in case the element doesn't
exist the first time. This can occur if the resource is the textarea has
not yet been rendered by the time the port is invoked.

Signed-off-by: Zoe Tian <ztian@pivotal.io>
Co-authored-by: Aidan Oldershaw <aoldershaw@pivotal.io>