web: behaviour: handle deltaMode for scrolling
concourse/concourse#5383

Consider `deltaMode` in the `wheel` event for determining how much to
scroll. Firefox is the only browser I've encountered that seems to
generate events with DeltaModeLine (when scrolling with a mouse wheel,
and not a trackpad).

According to https://bugzilla.mozilla.org/show_bug.cgi?id=943034, the
height of a line is based on the `font-size` of the scrollable element,
which in this case is 20px. The page height is set to 800px, which is
taken from
https://github.com/facebookarchive/fixed-data-table/blob/3a9bf338b22406169e7261f85ddeda22ddce3b6f/src/vendor_upstream/dom/normalizeWheel.js#L23.
Note that these values don't need to be exact - they should just feel
natural. I've never actually seen a DeltaModePage event in practice, so
I'm not sure if it will feel perfect, but oh well!

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>