web: behaviour: make icons consistently sized
concourse/concourse#4781

The pin icon was a fair bit smaller than the message and edit icons.
Additionally, after discussing with Matthew, we felt that the latter two icons
were larger than they should be.

This commit changes styling/padding to address these concerns, and make the
text more centered. It also adds padding to the bottom of the resource version
table.

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>