web: behaviour: convert scrollToId to a port
concourse/concourse#5454

With concourse/concourse#5275, I had attempted to fix scrolling to IDs
on the build page. While this sometimes worked, if it takes a long time
to render the log lines (common when there are many log lines), we would
fail to scroll since the ID did not yet exist on the page.

To fix this, revert the `scrollToId` behaviour to before
109dbcbaf623200632330bdb87d3c6e0a1c11182 i.e. use ports for this. Note
that the behaviour isn't exactly the same - rather than using
`scrollIntoView`, which moves the top of the element to the top of the
screen, set the `scrollTop` of the parent to `offsetTop` (minus some
padding so it's not right at the top). This is because the step has a
sticky header, so if we moved the log line right to the top, it would be
hidden behind this header.

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>