web: behaviour: add loading view to dashboard
As with concourse/concourse#5427, add a spinner to the dashboard before
any API responses come back. Note that this isn't as useful for 6.x
since typically the user will have a cached API response, which is very
quick to load. However, it's easy to do, so might as well!

This commit uses a slightly different design than
concourse/concourse#5427 - rather than having any text next to the
spinner, just display a spinner (in the center of the screen). I don't
have strong opinions either way, but the thinking was that we probably
don't need to tell the user what a spinner means! Happy to change it
to match the other design, though.

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>